requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.5
django-cors-headers==3.2.1
django-phonenumber-field==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
phonenumbers==8.12.3
Pillow==7.1.1
python-decouple==3.3
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
